refactor: split rpc.py into rpc_client.py and rpc_facade.py
