+++ rpent/utils/rpc/socket_rpc.py
@@ -32,7 +32,8 @@ import struct
 from typing import Any, Callable
 
 from rpent.utils.logging import get_logger
-from rpent.utils.rpc.rpc import RpcClient, check_response, make_error_response
+from rpent.utils.rpc.rpc_client import RpcClient, check_response
+from rpent.utils.rpc.rpc_facade import make_error_response
 
 logger = get_logger("socket_rpc")
